refactor: create utils.py file
